Work around an import race condition for strptime.
As discussed in http://bugs.python.org/issue7980, a multi-threaded program that
calls datetime.datetime.strptime in multiple threads can occasionally hit a
race condition that results in the (fatal) error "AttributeError: _strptime".

The workaround from the bug case is to call strptime at least once in the
parent thread before spawning, so this patch does that in what should be
a reliable way.